Process count and memory usage now reports using int64_t

+++ include/aws/common/system_info.h
@@ -112,16 +112,16 @@ int aws_get_cpu_usage(
     double *output);
 
 /**
- * Sets the passed double value as the memory usage in kilobytes. Only supported on Linux currently.
+ * Sets the passed int value as the memory usage in kilobytes. Only supported on Linux currently.
  */
 AWS_COMMON_API
-int aws_get_memory_usage(double *output);
+int64_t aws_get_memory_usage(int64_t *output);
 
 /**
- * Sets the passed double value as the process count. Only supported on Linux currently.
+ * Sets the passed int value as the process count. Only supported on Linux currently.
  */
 AWS_COMMON_API
-int aws_get_process_count(double *output);
+int64_t aws_get_process_count(int64_t *output);
 
 AWS_EXTERN_C_END
